pyproject.toml
[project]
name = "coolview"
version = "0.1.0"
description = "Show your images"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyside6>=6.10.1",
    "pysmb>=1.2.13",
    "webdavclient3>=3.14.6",
]

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true
